Log each call to behave in the decider.
We really need a unique id so that we can group together requests (and
possibly use it as the unique part of the logfilename). Also ideally,
we'd also log the contents of the behaviour so that each audit is a
self-contained record of the state at the time of writing.